read costs and rate data to be saved in corresponding thermal tables and trajectory

+++ src/main/java/com/rte_france/antares/datamanager_back/repository/model/ThermalCostEntity.java
@@ -22,7 +22,7 @@ public class ThermalCostEntity extends ThermalBaseEntity {
     private Double cost;
 
     @Column(name = "cost_year")
-    private Double year;
+    private Integer year;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "thermal_type_id")
